Adds a wiki macro/processor to show the entire dependency graph for the trac.
Pulls the list of all edges, and renders one big graph from there.
I'm building a string of the dot graph and then passing it to the
graphviz plugin's wiki processor, so this won't work without that
installed.  Has some undocumented arguments for controlling rendering
(colors and fontsize for now).

Renders each node as a link to the ticket with the summary in the
title.  Currently colors unblocked tickets and closed tickets.